fix(ava/insight): move augmented marks into augmentedMarks directory
