+++ packages/ava/src/insight/chart/index.ts
@@ -1,3 +1,3 @@
 export * from './generator/homogeneous';
 export * from './generator/insights';
-export * from './strategy';
+export * from './strategy/augmentedMarks';
