+++ packages/ava/src/insight/chart/strategy/augmentedMarks/correlation.ts
@@ -1,8 +1,7 @@
 import { Mark } from '@antv/g2';
 
-import { CorrelationInfo, InsightInfo } from '../../types';
-
-import { insight2ChartStrategy } from './chartStrategy';
+import { CorrelationInfo, InsightInfo } from '../../../types';
+import { insight2ChartStrategy } from '../chart';
 
 // eslint-disable-next-line @typescript-eslint/no-unused-vars
 export const correlationStrategy = (insight: InsightInfo<CorrelationInfo>, patterns: CorrelationInfo[]): Mark[] => {
