+++ packages/ava/src/insight/chart/strategy/augmentedMarks/homogeneous.ts
@@ -1,6 +1,6 @@
 import { G2Spec } from '@antv/g2';
 
-import { HomogeneousPatternInfo, InsightInfo } from '../../types';
+import { HomogeneousPatternInfo, InsightInfo } from '../../../types';
 
 // eslint-disable-next-line @typescript-eslint/no-unused-vars
 export const homogeneousStrategy = (insight: InsightInfo<HomogeneousPatternInfo>): G2Spec => {
