+++ packages/ava/src/insight/chart/strategy/augmentedMarks/index.ts
@@ -1,0 +1,8 @@
+export { trendStrategy, trendAugmentedMarksStrategy } from './trend';
+export { lowVarianceStrategy, lowVarianceAugmentedMarkStrategy } from './lowVariance';
+export { categoryOutlierStrategy, categoryOutlierAugmentedMarksStrategy } from './categoryOutlier';
+export { changePointStrategy, changePointAugmentedMarksStrategy } from './changePoint';
+export { majorityStrategy } from './majority';
+export { correlationStrategy } from './correlation';
+export { homogeneousStrategy } from './homogeneous';
+export { timeSeriesOutlierStrategy } from './timeSeriesOutlier';
