+++ packages/ava/src/insight/chart/strategy/augmentedMarks/majority.ts
@@ -1,8 +1,7 @@
 import { Mark } from '@antv/g2';
 
-import { MajorityInfo, InsightInfo } from '../../types';
-
-import { insight2ChartStrategy } from './chartStrategy';
+import { MajorityInfo, InsightInfo } from '../../../types';
+import { insight2ChartStrategy } from '../chart';
 
 // eslint-disable-next-line @typescript-eslint/no-unused-vars
 export const majorityStrategy = (insight: InsightInfo<MajorityInfo>, patterns: MajorityInfo[]): Mark[] => {
