+++ packages/ava/src/insight/chart/strategy/chart.ts
