+++ packages/ava/src/insight/chart/strategy/commonMarks/index.ts
@@ -2,3 +2,4 @@ export { lineMarkStrategy } from './lineMark';
 export { pointMarkStrategy } from './pointMark';
 export { textMarkStrategy } from './textMark';
 export { areaMarkStrategy } from './areaMark';
+export { transparent } from './transparent';
