+++ packages/ava/src/insight/chart/strategy/index.ts
@@ -1,8 +1,4 @@
-export { trendStrategy, trendAugmentedMarksStrategy } from './trend';
-export { lowVarianceStrategy, lowVarianceAugmentedMarkStrategy } from './lowVariance';
-export { categoryOutlierStrategy, categoryOutlierAugmentedMarksStrategy } from './categoryOutlier';
-export { changePointStrategy, changePointAugmentedMarksStrategy } from './changePoint';
-export { majorityStrategy } from './majority';
-export { correlationStrategy } from './correlation';
-export { homogeneousStrategy } from './homogeneous';
-export { timeSeriesOutlierStrategy } from './timeSeriesOutlier';
+export { insight2ChartStrategy } from './chart';
+export { viewSpecStrategy } from './viewSpec';
+export * from './commonMarks';
+export * from './augmentedMarks';
