+++ packages/ava/src/insight/chart/strategy/viewSpec.ts
@@ -1,6 +1,6 @@
 import { G2Spec, Mark } from '@antv/g2';
 
-export const wrapViewSpec = (marks: Mark[]): G2Spec => {
+export const viewSpecStrategy = (marks: Mark[]): G2Spec => {
   return {
     type: 'view',
     theme: 'classic',
